Split outside_connections.rs (1779 lines) into sub-module directory
Break the monolithic outside_connections.rs into logical sub-files:
- types.rs (118 lines): ConnectionType, OutsideConnection, OutsideConnections, ConnectionStat
- effects.rs (64 lines): ConnectionEffects computation
- detection.rs (194 lines): edge helpers + highway/railway/seaport/airport detection
- systems.rs (152 lines): update_outside_connections Bevy system
- tests_types.rs, tests_detection.rs, tests_effects.rs, tests_application.rs: split tests
- mod.rs (32 lines): plugin + re-exports

All pub items are re-exported from mod.rs so callers need no import changes.

Part of #1389
